requirements.txt
asgiref==3.2.5
bcrypt==3.1.7
cffi==1.14.0
Django==3.0
djangorestframework==3.11.0
gunicorn==19.9.0
mysqlclient==1.4.6
Pillow==7.0.0
pycparser==2.20
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
whitenoise==3.3.1
